feat: Add LSP and VSCode Extension for YINI
This commit introduces comprehensive IDE support for the YINI language by adding a Language Server Protocol (LSP) server and a VSCode extension.

The new IDE support includes the following features:
- **Syntax Highlighting:** Full-color syntax highlighting for all `.yini` files is provided via a new TextMate grammar.
- **Real-time Diagnostics:** The LSP server uses the core YINI library to parse files on the fly and report syntax errors directly in the editor, providing immediate feedback to the user.
- **Code Completion:** The server provides intelligent code completion for macros, suggesting defined `@` variables to improve developer productivity.

To enable this, the following changes were made:
- A new C# project for the LSP server (`ide/LspServer`) was created using the `OmniSharp.Extensions.LanguageServer` library.
- A new VSCode extension project (`ide/VSIX`) was created to manage the language client, syntax grammar, and other IDE-specific configurations.
- The core YINI C++ library was enhanced with a `load_from_string` method to allow the LSP to parse in-memory document content.
- The C-API and C# wrapper were extended to expose the new functionality and the list of defined macros for code completion.
- The unified CMake build system was updated to build the new LSP server project.
- The `README.md` file was updated with instructions on how to build and use the new VSCode extension.

+++ src/Lexer/Lexer.cpp
@@ -127,7 +127,7 @@ namespace YINI
 
     bool Lexer::isAtEnd()
     {
-        return m_current >= m_source.length();
+        return static_cast<size_t>(m_current) >= m_source.length();
     }
 
     void Lexer::addToken(TokenType type)
